+++ example/health_check_server.py
@@ -6,7 +6,7 @@ import asyncio
 import http
 import websockets
 
-def health_check(path, request_headers):
+async def health_check(path, request_headers):
     if path == '/health/':
         return http.HTTPStatus.OK, [], b'OK\n'